+++ packages/dara-core/dara/core/js_tooling/custom_js_scaffold/index.tsx
@@ -1,1 +1,0 @@
-export { default as LocalJsComponent } from './local-js-component';